move upgrade version constant file
